+++ modules/terraform-zscc-network-gcp/variables.tf
@@ -90,6 +90,18 @@ variable "fw_cc_service_default_name" {
   default     = null
 }
 
+variable "fw_cc_mgmt_zssupport_tunnel_name" {
+  type        = string
+  description = "The name of the compute firewall created on the user defined Cloud Connector Management VPC Network permitting CC to establish zssupport tunnel"
+  default     = null
+}
+
+variable "support_access_enabled" {
+  type        = bool
+  description = "Enable a specific outbound firewall rule for Cloud Connector to be able to establish connectivity for Zscaler support access. Default is true"
+  default     = true
+}
+
 
 # BYO (Bring-your-own) variables list